backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
face_recognition==1.3.0
opencv-python==4.10.0.84
pymongo==4.8.0
Pillow==10.4.0
python-dotenv==1.0.1
numpy==1.26.4
